Upgraded RealmAppImport to use the realm-cli

+++ packages/realm-app-importer/src/RealmAppImporter.ts
@@ -27,15 +27,16 @@ export interface RealmAppImporterOptions {
     baseUrl: string;
     username: string;
     password: string;
-    stitchConfigPath: string;
+    realmConfigPath: string;
     appsDirectoryPath: string;
+    cleanUp?: boolean;
 }
 
 export class RealmAppImporter {
     private readonly baseUrl: string;
     private readonly username: string;
     private readonly password: string;
-    private readonly stitchConfigPath: string;
+    private readonly realmConfigPath: string;
     private readonly appsDirectoryPath: string;
 
     private accessToken: string | undefined;
@@ -44,34 +45,37 @@ export class RealmAppImporter {
         baseUrl,
         username,
         password,
-        stitchConfigPath,
+        realmConfigPath,
         appsDirectoryPath,
+        cleanUp = true,
     }: RealmAppImporterOptions) {
         this.baseUrl = baseUrl;
         this.username = username;
         this.password = password;
-        this.stitchConfigPath = stitchConfigPath;
+        this.realmConfigPath = realmConfigPath;
         this.appsDirectoryPath = appsDirectoryPath;
 
-        process.on("exit", () => {
-            // Remove any stitch configuration
-            if (fs.existsSync(this.stitchConfigPath)) {
-                console.log(`Deleting ${this.stitchConfigPath}`);
-                fs.removeSync(this.stitchConfigPath);
-            }
-            // If there is nothing the the apps directory, lets delete it
-            if (fs.existsSync(this.appsDirectoryPath)) {
-                console.log(`Deleting ${this.appsDirectoryPath}`);
-                fs.removeSync(this.appsDirectoryPath);
-            }
-        });
+        if (cleanUp) {
+            process.on("exit", () => {
+                // Remove any stitch configuration
+                if (fs.existsSync(this.realmConfigPath)) {
+                    console.log(`Deleting ${this.realmConfigPath}`);
+                    fs.removeSync(this.realmConfigPath);
+                }
+                // If there is nothing the the apps directory, lets delete it
+                if (fs.existsSync(this.appsDirectoryPath)) {
+                    console.log(`Deleting ${this.appsDirectoryPath}`);
+                    fs.removeSync(this.appsDirectoryPath);
+                }
+            });
+        }
     }
 
     public async importApp(
         appTemplatePath: string,
     ): Promise<{ appId: string }> {
-        const { name: appName } = this.loadStichJson(appTemplatePath);
-        await this.login();
+        const { name: appName } = this.loadAppConfigJson(appTemplatePath);
+        await this.logIn();
         const groupId = await this.getGroupId();
 
         // Get or create an application
@@ -96,10 +100,10 @@ export class RealmAppImporter {
         this.copyAppTemplate(appPath, appTemplatePath);
 
         // Import
-        this.stitchCli(
+        this.realmCli(
             "import",
             "--config-path",
-            this.stitchConfigPath,
+            this.realmConfigPath,
             "--base-url",
             this.baseUrl,
             "--app-name",
@@ -119,6 +123,10 @@ export class RealmAppImporter {
         return { appId };
     }
 
+    private get apiUrl() {
+        return `${this.baseUrl}/api/admin/v3.0`;
+    }
+
     private loadJson(filePath: string) {
         try {
             const content = fs.readFileSync(filePath, "utf8");
@@ -130,15 +138,15 @@ export class RealmAppImporter {
         }
     }
 
-    private loadStichJson(appTemplatePath: string) {
-        const stitchJsonPath = path.resolve(appTemplatePath, "stitch.json");
-        return this.loadJson(stitchJsonPath);
+    private loadAppConfigJson(appTemplatePath: string) {
+        const configJsonPath = path.resolve(appTemplatePath, "config.json");
+        return this.loadJson(configJsonPath);
     }
 
     private loadSecretsJson(appTemplatePath: string) {
-        const stitchJsonPath = path.resolve(appTemplatePath, "secrets.json");
-        if (fs.existsSync(stitchJsonPath)) {
-            return this.loadJson(stitchJsonPath);
+        const secretsJsonPath = path.resolve(appTemplatePath, "secrets.json");
+        if (fs.existsSync(secretsJsonPath)) {
+            return this.loadJson(secretsJsonPath);
         } else {
             return {};
         }
@@ -154,20 +162,19 @@ export class RealmAppImporter {
         }
     }
 
-    private stitchCli(...args: string[]) {
-        cp.spawnSync("stitch-cli", args, { stdio: "inherit" });
+    private realmCli(...args: string[]) {
+        cp.spawnSync("realm-cli", args, { stdio: "inherit" });
     }
 
-    private async login() {
-        const url = `${this.baseUrl}/api/admin/v3.0/auth/providers/local-userpass/login`;
-        const body = JSON.stringify({
-            username: this.username,
-            password: this.password,
-        });
+    private async logIn() {
+        const url = `${this.apiUrl}/auth/providers/local-userpass/login`;
         const response = await fetch(url, {
             method: "POST",
             headers: { "content-type": "application/json" },
-            body,
+            body: JSON.stringify({
+                username: this.username,
+                password: this.password,
+            }),
         });
         // Store the access and refresh tokens
         const responseBody = await response.json();
@@ -186,16 +193,12 @@ export class RealmAppImporter {
         refreshToken: string,
         accessToken: string,
     ) {
-        const stitchConfig = [
+        const realmConfig = [
             `public_api_key: ${username}`,
             `refresh_token: ${refreshToken}`,
             `access_token: ${accessToken}`,
         ];
-        fs.writeFileSync(
-            this.stitchConfigPath,
-            stitchConfig.join("\n"),
-            "utf8",
-        );
+        fs.writeFileSync(this.realmConfigPath, realmConfig.join("\n"), "utf8");
     }
 
     private async getProfile() {
